Extract BrailleString class to other file
